* Fix #973988: levels only show once for multiple audio tracks on one instr * Fix #991512: matrix IPB doesn't work for soft synths
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@6155 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/trackbuttons.h
@@ -185,8 +185,8 @@ public slots:
     void slotToggleMutedTrack(int mutedTrack);
     void slotUpdateTracks();
     void slotRenameTrack(QString newName, Rosegarden::TrackId trackId);
-    void slotSetTrackMeter(double value, int position);
-    void slotSetMetersByInstrument(double value, Rosegarden::InstrumentId id);
+    void slotSetTrackMeter(float value, int position);
+    void slotSetMetersByInstrument(float value, Rosegarden::InstrumentId id);
 
     void slotInstrumentSelection(int);
     void slotInstrumentPopupActivated(int);
